Refactor so invert also has discovery entry point
