+++ packages/backend/src/config/Config.ts
@@ -15,10 +15,8 @@ export interface Config {
   readonly database: DatabaseConfig | false
   readonly api: ApiConfig | false
   readonly health: HealthConfig
-  readonly invert: InversionConfig | false
   readonly tvl: TvlConfig | false
   readonly activity: ActivityConfig | false
-  readonly discovery: DiscoveryModuleConfig | false
   readonly discoveryWatcher: DiscoveryWatcherConfig | false
 }
 
@@ -63,25 +61,12 @@ export interface HealthConfig {
   readonly startedAt: string
   readonly commitSha: string
 }
-
-export interface InversionConfig {
-  readonly file: string
-  readonly useMermaidMarkup: boolean
-}
-
 export interface ActivityConfig {
   readonly starkexApiKey: string
   readonly starkexCallsPerMinute: number
   readonly skipExplicitExclusion: boolean
   readonly allowedProjectIds?: string[]
   readonly projects: Record<string, Layer2TransactionApi | undefined>
-}
-
-export interface DiscoveryModuleConfig {
-  readonly project: string
-  readonly blockNumber?: number
-  readonly alchemyApiKey: string
-  readonly etherscanApiKey: string
 }
 
 export interface MetricsAuthConfig {
