+++ packages/backend/src/core/discovery/runDiscovery.ts
@@ -1,7 +1,7 @@
 import { MainnetEtherscanClient } from '@l2beat/shared'
 import { providers } from 'ethers'
 
-import { DiscoveryModuleConfig } from '../../config/Config'
+import { DiscoveryModuleConfig } from '../../config/config.discovery'
 import { ConfigReader } from './ConfigReader'
 import { discover } from './discover'
 import { DiscoveryLogger } from './DiscoveryLogger'
